+++ classes/Keys.js
@@ -2,15 +2,10 @@
 'use strict'
 
 const OneBlinkAPI = require('../lib/one-blink-api.js')
-const getTenantUrl = require('../lib/tenant')
 module.exports = class Keys extends OneBlinkAPI {
   constructor(options /* : ConstructorOptions */) {
     options = options || {}
-    super(
-      getTenantUrl(options.tenant, options.oneBlinkAPIOrigin),
-      options.accessKey,
-      options.secretKey,
-    )
+    super(options.accessKey, options.secretKey, options.tenant)
   }
 
   getKey(keyId /* : ?mixed */) /* : Promise<Key> */ {