+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- verifyJWT function in FormsApps Class
+
 # 0.6.1 (2020-06-03)
 
 ### Added